HID: ft260: skip unexpected HID input reports
The FT260 does not send unsolicited reports. However, in theory, they can
be emitted by a specially crafted malicious USB device masquerading as FT260
when the attacker has physical access to the USB port. In this case, the
read_buf pointer poinits to the final data portion of the previous I2C Read
transfer, and the memcpy invoked in the ft260_raw_event() will try copying
the content of the unexpected report into the wrong location.

This commit sets the Read buffer pointer to NULL on the I2C Read transaction
completion and checks it in the ft260_raw_event() to detect and skip the
unsolicited Input report.

Signed-off-by: Michael Zaidman <michael.zaidman@gmail.com>