revert to pre missed nack on big reads commit
The intention of the "HID: FT260: missed NACK on big i2c reads" commit was
to separate the big read to two i2c read requests, first of 60 bytes with
FT260_FLAG_START_STOP and second - the rest of the data length. In that
way, the controller correctly returns NACK on the first i2c read from a
non-existing device.

The "HID: ft260: wake up device from power saving mode" commit provides
a better solution. Hence, revert the former to apply the latter.

Signed-off-by: Michael Zaidman <michael.zaidman@gmail.com>