hid-ft260: reduce bus error message severity
The current driver prints an error message into the kernel log upon
accessing a nonexistent device. The assumption was that the application
code usually does not try to access nonexistent I2C devices. So, the
driver writes an error message into the dmesg log if such access happens,
pointing to the possible issue. On the other hand,  there are use cases
besides the i2cdetect that may scan the I2c bus to discover the devices
such as HWMON class devices or Transceivers modules Monitoring. In these
cases, the I2C bus error reported by the device is not an error, and
should not be reported.

Signed-off-by: Michael Zaidman <michael.zaidman@gmail.com>